Add Converters, Models, Resources and Services to App

+++ Models/WoundClassification.cs
@@ -2,7 +2,7 @@
 public class WoundClassification
 {
     public WoundType PrimaryType { get; set; }
-    public double Confidence { get; set; }
-    public Dictionary<WoundType, double> AlternativeTypes { get; set; } = new();
+    public float Confidence { get; set; }
+    public Dictionary<WoundType, float> AlternativeTypes { get; set; } = new();
     public string? ErrorMessage { get; set; }
 }
